+++ src/main/java/pe/edu/vallegrande/database/repository/BasicServiceRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.vallegrande.database.model.BasicService;
 
 @Repository
-public interface BasicServiceRepository extends ReactiveCrudRepository<BasicService, Long> {
+public interface BasicServiceRepository extends ReactiveCrudRepository<BasicService, Integer> {
      
 }
