+++ src/main/java/pe/edu/vallegrande/database/repository/FamilyHealthRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 import pe.edu.vallegrande.database.model.FamilyHealth;
 
-public interface FamilyHealthRepository extends ReactiveCrudRepository<FamilyHealth, Long> {
+public interface FamilyHealthRepository extends ReactiveCrudRepository<FamilyHealth, Integer> {
     
 }
