+++ src/main/java/pe/edu/vallegrande/database/repository/HousingEnvironmentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 import pe.edu.vallegrande.database.model.HousingEnvironment;
 
-public interface HousingEnvironmentRepository extends ReactiveCrudRepository<HousingEnvironment,Long> {
+public interface HousingEnvironmentRepository extends ReactiveCrudRepository<HousingEnvironment, Integer> {
     
 }
