+++ src/main/java/pe/edu/vallegrande/database/repository/LaborAutonomyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 import pe.edu.vallegrande.database.model.LaborAutonomy;
 
-public interface LaborAutonomyRepository extends ReactiveCrudRepository<LaborAutonomy,Long> {
+public interface LaborAutonomyRepository extends ReactiveCrudRepository<LaborAutonomy, Integer> {
     
 }
