+++ src/main/java/pe/edu/vallegrande/database/repository/SocialLifeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
 import pe.edu.vallegrande.database.model.SocialLife;
 
-public interface SocialLifeRepository extends ReactiveCrudRepository<SocialLife,Long> {
+public interface SocialLifeRepository extends ReactiveCrudRepository<SocialLife, Integer> {
     
 }
